+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Breaking changes:
 * Replaced `user_id` parameter of `r0::membership::invite_user` with `recipient` 
   to allow invitation of users by either Matrix or third party identifiers.
 * Remove deprecated endpoint `r0::contact::create_contact` (deprecated in r0.6.0)
+* Add lazy-loading options to `r0::filter::RoomEventFilter` (introduced in r0.5.0)
 
 # 0.6.0